requirements.txt
scikit-image
pillow
shiny